Permet d'importer le script avec l'utilisation iframe-integration.js
