+++ iframes/iframe-integration.js
@@ -1,0 +1,28 @@
+import { iframeResize } from "iframe-resizer"
+
+const script = document.currentScript
+const src = new URL(`${process.env.BASE_URL}/simulation`)
+
+iframeResize({ log: false }, "#iframe")
+
+src.searchParams.set("iframe", true)
+src.searchParams.set(
+  "integratorUrl",
+  encodeURIComponent(window.location.href.toString())
+)
+
+const iframe = document.createElement("iframe")
+const iframeAttributes = {
+  id: "simulateur",
+  src: src.toString(),
+  style: "border: none; width: 100%; display: block; height: 700px",
+  allow: "clipboard-write",
+  allowfullscreen: true,
+  webkitallowfullscreen: true,
+  mozallowfullscreen: true,
+}
+for (var key in iframeAttributes) {
+  iframe.setAttribute(key, iframeAttributes[key])
+}
+
+script.before(iframe)
